Added AssetSummary type to avoid unmarshalling problems.
The screenly api is not strict in terms of the types of the JSON fields
that it accepts and returns, eg. Duration can be given as a string or an
integer. In particular there is a problem with Boolean values - if they
haven't been specified at all then they default to an empty string - and
this causes problems when trying to unmarshal to the Asset type which
must specify the types of its members. The AssetSummary is just an Asset
with any problematic fields removed, so the json unmarshaller will
ignore them.